# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Deloitte Initiative Discovery Platform",
    "description": "\n    ## Deloitte Initiative Discovery Platform API\n    \n    A unified platform for Deloitte analysts to discover and engage with firm initiatives.\n    \n    ### Features:\n    * **Authentication**: Azure AD B2C SSO integration\n    * **User Profiles**: Manage analyst and leader profiles\n    * **Initiative Management**: Create, update, and manage initiatives\n    * **Discovery Tools**: Search, filter, and browse initiatives\n    * **AI Recommendations**: Personalized initiative matching\n    * **Engagement**: Save, bookmark, and track initiatives\n    \n    ### User Roles:\n    * **Analyst**: Browse, search, and receive recommendations\n    * **Leader**: Create and manage initiatives\n    ",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/initiatives/": {
      "post": {
        "tags": [
          "Initiatives"
        ],
        "summary": "Create initiative",
        "description": "Create a new initiative.\n\nRequires leader role.\n\nThe initiative will automatically:\n- Generate AI tags based on description (future enhancement)\n- Create vector embeddings for semantic search (future enhancement)",
        "operationId": "create_initiative_api_v1_initiatives__post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InitiativeCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InitiativeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Initiatives"
        ],
        "summary": "List initiatives",
        "description": "List initiatives with optional filtering.\n\nSupports filtering by:\n- Status (open, active, full, closed)\n- Practice area\n\nPagination via skip and limit parameters.",
        "operationId": "list_initiatives_api_v1_initiatives__get",
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/InitiativeStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "practice_area",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Practice Area"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InitiativeList"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/initiatives/{initiative_id}": {
      "get": {
        "tags": [
          "Initiatives"
        ],
        "summary": "Get initiative by ID",
        "description": "Get a specific initiative by ID.\n\nIncrements view count if user is authenticated.",
        "operationId": "get_initiative_api_v1_initiatives__initiative_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "initiative_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Initiative Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InitiativeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Initiatives"
        ],
        "summary": "Update initiative",
        "description": "Update an initiative.\n\nOnly the initiative owner or admin can update.",
        "operationId": "update_initiative_api_v1_initiatives__initiative_id__put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "initiative_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Initiative Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InitiativeUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InitiativeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Initiatives"
        ],
        "summary": "Delete initiative",
        "description": "Delete an initiative.\n\nOnly the initiative owner or admin can delete.",
        "operationId": "delete_initiative_api_v1_initiatives__initiative_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "initiative_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Initiative Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/initiatives/my/initiatives": {
      "get": {
        "tags": [
          "Initiatives"
        ],
        "summary": "Get my initiatives",
        "description": "Get all initiatives created by the current user.",
        "operationId": "get_my_initiatives_api_v1_initiatives_my_initiatives_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/InitiativeResponse"
                  },
                  "type": "array",
                  "title": "Response Get My Initiatives Api V1 Initiatives My Initiatives Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "InitiativeCreate": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "practice_area": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Practice Area"
          },
          "skills_needed": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Skills Needed"
          },
          "industries": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Industries"
          },
          "time_commitment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Time Commitment"
          },
          "duration": {
            "allOf": [
              {
                "$ref": "#/components/schemas/InitiativeDuration"
              }
            ],
            "default": "ongoing"
          },
          "duration_details": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Details"
          },
          "role_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Role Type"
          },
          "contact_person": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contact Person"
          },
          "contact_email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contact Email"
          }
        },
        "type": "object",
        "required": [
          "title",
          "description"
        ],
        "title": "InitiativeCreate"
      },
      "InitiativeDuration": {
        "type": "string",
        "enum": [
          "short_term",
          "ongoing",
          "fixed_duration"
        ],
        "title": "InitiativeDuration"
      },
      "InitiativeList": {
        "properties": {
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/InitiativeResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          }
        },
        "type": "object",
        "required": [
          "total",
          "items",
          "page",
          "page_size"
        ],
        "title": "InitiativeList"
      },
      "InitiativeResponse": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "practice_area": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Practice Area"
          },
          "skills_needed": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Skills Needed"
          },
          "industries": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Industries"
          },
          "time_commitment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Time Commitment"
          },
          "duration": {
            "allOf": [
              {
                "$ref": "#/components/schemas/InitiativeDuration"
              }
            ],
            "default": "ongoing"
          },
          "duration_details": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Details"
          },
          "role_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Role Type"
          },
          "contact_person": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contact Person"
          },
          "contact_email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contact Email"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "status": {
            "$ref": "#/components/schemas/InitiativeStatus"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tags"
          },
          "owner_id": {
            "type": "integer",
            "title": "Owner Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "view_count": {
            "type": "integer",
            "title": "View Count",
            "default": 0
          },
          "save_count": {
            "type": "integer",
            "title": "Save Count",
            "default": 0
          },
          "application_count": {
            "type": "integer",
            "title": "Application Count",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "title",
          "description",
          "id",
          "status",
          "owner_id",
          "created_at",
          "updated_at"
        ],
        "title": "InitiativeResponse"
      },
      "InitiativeStatus": {
        "type": "string",
        "enum": [
          "open",
          "active",
          "full",
          "closed"
        ],
        "title": "InitiativeStatus"
      },
      "InitiativeUpdate": {
        "properties": {
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "practice_area": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Practice Area"
          },
          "skills_needed": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Skills Needed"
          },
          "industries": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Industries"
          },
          "time_commitment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Time Commitment"
          },
          "duration": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/InitiativeDuration"
              },
              {
                "type": "null"
              }
            ]
          },
          "duration_details": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Details"
          },
          "role_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Role Type"
          },
          "contact_person": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contact Person"
          },
          "contact_email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contact Email"
          },
          "status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/InitiativeStatus"
              },
              {
                "type": "null"
              }
            ]
          },
          "tags": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tags"
          }
        },
        "type": "object",
        "title": "InitiativeUpdate"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}